add base panels
The layer “addition_base” includes metal panels to be welded to the
bottom of the floor supports in “floor_support” layer creating a new
base under the floor. The enclosed cavities created between each floor
support will be the space for the under-floor heating. I imagine the
next layer, immediately above these new panels and on the sides of the
floor supports, to be an insulator, like aerogel, then, within that,
the PEX tubing carrying the heat exchange liquid in whatever setting it
needs to be in (e.g., concrete, etcetera).